test(09-cr-tests): multiply pages by the page size before dividing
CodeRabbit CLI finding on the tests chunk (minor): residentKb() computed
resident * (page / 1024), which discards the fraction a page size below
1024 leaves and keeps the product in a long, which a 32-bit long
overflows past two million pages. Now (long long)resident * page / 1024.
The three binaries that read it were rebuilt and run: wm2_config_smoke
766 assertions in 71 cases, wm_runtime 467 in 39, wm_resource passed.

+++ tests/support/WmFixture.h
@@ -993,7 +993,10 @@ inline bool residentKb(pid_t pid, long& out)
     // against a live process. A silently-zero reading is worse than a false.
     const long page = ::sysconf(_SC_PAGESIZE);
     if (page <= 0) return false;
-    out = resident * (page / 1024);
+    // Multiply first, in 64 bits: `page / 1024` would discard the fraction of
+    // a page a size below 1024 leaves, and `resident * page` in a 32-bit long
+    // overflows past two million pages.
+    out = static_cast<long>((static_cast<long long>(resident) * page) / 1024);
     return true;
 }
